+++ src/Valkyrja/Http/Message/Stream/Factory/PsrStreamFactory.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Valkyrja Framework package.
+ *
+ * (c) Melech Mizrachi <melechmizrachi@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Valkyrja\Http\Message\Stream\Factory;
+
+use Psr\Http\Message\StreamInterface;
+use Valkyrja\Http\Message\Stream\Contract\StreamContract;
+use Valkyrja\Http\Message\Stream\Enum\PhpWrapper;
+use Valkyrja\Http\Message\Stream\Stream;
+
+abstract class PsrStreamFactory
+{
+    /**
+     * Get a Valkyrja Stream object from a PSR StreamInterface object.
+     */
+    public static function fromPsr(StreamInterface $stream): StreamContract
+    {
+        $stream->rewind();
+        $contents = $stream->getContents();
+        $stream->rewind();
+
+        $valkyrjaStream = new Stream(PhpWrapper::temp);
+        $valkyrjaStream->write($contents);
+        $valkyrjaStream->rewind();
+
+        return $valkyrjaStream;
+    }
+}
